pyproject.toml
[project]
name = "patient-tool"
version = "1.6"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flet[all]>=0.80.2",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "pyinstaller>=6.11.1",
    "fpdf2>=2.8.5",
    "pydantic",
]
